Fix project parameter in App Identity README.
The gcloud command was getting the project argument when I was passing
it in as -Dgcloud_project. After changing to -Dgcloud.gcloud_project it
now deploys to the project I specify there. I hadn't noticed this until
I had modified my gcloud config and deployments started going to the
wrong project.