Support App Engine or Jetty for local testing.
This script will allow us to run tests against a local server for
testing. Since the App Engine and Jetty server running and testing is so
similar, I combined the scripts rather than copying. The only
differences are which plugin to use and what string to wait for before
running curl.

Since the bookshelf sample also needs a `-Plocal` profile argument, I
extended the script to allow passing through arbitrary arguments to
Maven.