Cloud Spanner: Allow version overrides on the command line.
Right now, the Cloud Spanner docs go stale whenever we upgrade the parent to a new version. The docs are stale right now, in fact; they say `1.0.10`, but the current version is `1.0.11` (example: [1]).

With this change, we can update the docs to tell customers to run something like `mvn package -Drevision=0.1.0-SNAPSHOT`, and then use `0.1.0-SNAPSHOT` throughout the examples.

[1]: https://cloud.google.com/spanner/docs/getting-started/java/