Update Firestore Quickstart to use exec:exec
When exec-maven-plugin runs the `exec:java` goal, it executes the java
program in the maven process (there currently isn't an option to fork
instead of running it in process). As a side effect of this, shutdown
hooks are not run when the program has completed running. In the case
of our Quickstart that means the daemon thread for grpc and gax do not
shutdown when the program has completed running, resulting in maven
listing all the threads that are still running and showing a stacktrace.

This change uses `exec:exec` and manually builds the java command to be
run, thereby forcing the program to be ran in a forked process and able
to use the normal shutdown process.