Upgrade dependency version for Appengine Java8 Endpoints sample (#1603)
* Upgrade the version of Endpoints Management Framework to newly released
1.0.12.

* Update Gradle to 4.10.3 and Endpoints Frameworks to the current version.

build.gradle referenced an old version of Endpoints Frameworks. This CL
updates the selectors to the current releases of Endpoints Frameworks
and Endpoints Management.

Transitively, Gradle version is upgrated to 4.10.6, because the new
Frameworks releases use advanced POM features, support for which is
present only in gradle 4.6+

* Update gradle endpoints-framework-gradle-plugin.

The new uses less of the deprecated Gradle features. All of the features
used by the new version still work in Gradle 5.6.2