fix: only run test if restore succeeded
The database restore operation could be given up by the test if it fails
too many times with an error indicating that too many restore operations
are running. The test for listing database operations would still be
executed and would then look for the optimize operation of the previous
restore. This operation would not be found, as the restore operation was
skipped.

Fixes #3541