Specify member type (#3602)
Saying "member:user@example.com" isn't helpful, since the prefix needs to be "user", "serviceAccount", or "group", not "member'"

Since this is a quickstart, we should just pick one type and stick with that. "user" is a good way to demonstrate core functionality, so I picked that type.

lesv: using ADMIN to push as the bug is somewhere else. See #3605 for tracking issue