Fix and add logging to com.example.spanner.changestreams.ChangeStreamSampleIT: testChangeStreamSample
In my local testing, sometimes the test failed with timeout executing DDL, so I increased the timeout from 2 minutes to 15 minutes. After the fix, I run the test 100 times locally and all of them passed.

I am not sure if this will fix the problem since the test was failed at a different place in Sponge, but I cannot repro it locally and I cannot see the logging since the standard output was hidden, so I also print the standard output and hopefully we will see some logging to debug if it fails another time.