cdn/signed-urls: use correct HMac algorithm
The Cloud CDN Signed URLs and Signed Cookies features expect signatures
to be created with the HMAC-SHA1 algorithm; Google's servers validate
signatures using this algorithm. PRs #8103, #8136, and #8183 updated the
Java code samples to use HMAC-SHA256 instead, based on the argument that
according to GCP's KMS docs on choosing a good algorithm, HMAC-SHA256 is
recommended.

While it may be a better algorithm to use when designing a new system,
it is not the algorithm used by the Signed URLs and Signed Cookies
feature, so these code samples are currently 100% wrong and do not
actually work to implement URL/cookie signing! (Fortunately the non-Java
code samples are still correct.)

This PR reverts changes made in the above PRs to use the correct
HMAC-SHA1 algorithm.