feat(security-command-center): add sample for creating client with regional endpoint
The sample also shows how to create a client for the default endpoint. That's because SCC's regional endpoints support only a limited subset of SCC resources that are subject to data residency controls: https://cloud.google.com/security-command-center/docs/data-residency-support#scc-regional-urls-rest

As a result, it's likely that customers will need to create two clients: one for their regional endpoint, to call the methods it supports, and another for the default endpoint, to call all other methods.